requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
djangorestframework-simplejwt==5.3.0
Pillow>=10.0.0
python-decouple==3.8





